Move databrowser_api and data_portal to freva_rest
