+++ freva-rest/src/freva_rest/databrowser_api/mock/files.json
