+++ freva-rest/src/freva_rest/databrowser_api/mock/latest.json
