+++ freva-rest/src/freva_rest/databrowser_api/mock_broken/bears.nc
