+++ freva-rest/src/freva_rest/databrowser_api/schema.py
